+++ CHANGELOG.md
@@ -9,6 +9,8 @@ inside pull requests and have no tag.
 
 ## [Unreleased]
 
+## [0.5.5] - 2026-09-11
+
 ### Added
 
 - A backend or data-ml repository's ruleset requires the check `image`, the